+++ app/controllers/auth.js
@@ -1,0 +1,18 @@
+var express = require('express'),
+  router = express.Router(),
+  passport = require('passport'),
+  db = require('../models');
+
+  router.post('/signup', passport.authenticate('local-signup', {
+        successRedirect : '/profile', // redirect to the secure profile section
+        failureRedirect : '/signup', // redirect back to the signup page if there is an error
+        failureFlash : true // allow flash messages
+    }));
+
+  router.post('/login', passport.authenticate('local-login', {
+      successRedirect : '/profile', // redirect to the secure profile section
+      failureRedirect : '/login', // redirect back to the signup page if there is an error
+      failureFlash : true // allow flash messages
+  }));
+
+  module.exports = router;
