+++ config/passport.js
@@ -27,8 +27,8 @@ module.exports = function(passport) {
       passwordField : 'password',
       passReqToCallback : true
   },
-  function(req, email, password, done) {
 
+  function(req, email, password, done) {
   process.nextTick(function() {
 
     User.findOne({ where: { email :  email } })
